2009-8-6: maximum age veranderd

+++ Dynamo-HIA_UI/src/nl/rivm/emi/dynamo/data/types/atomic/MaxAge.java
@@ -7,6 +7,6 @@ public class MaxAge extends AbstractAge implements PayloadType<Integer> {
 	static final protected String XMLElementName = "maxAge";
 
 	public MaxAge() {
-		super(XMLElementName, new Integer(1), AbstractAge.MAXAGE);
+		super(XMLElementName, new Integer(0), AbstractAge.MAXAGE);
 	}
 }
